# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: M&E Garments storefront and shopping-assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /account/profile:
    get:
      operationId: getCustomerProfile
      tags: [account]
      summary: Get the authenticated customer's profile
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateCustomerProfile
      tags: [account]
      summary: Update the authenticated customer's profile
      security: [{ ClerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerProfileInput" }
      responses:
        "200":
          description: Updated customer profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerProfile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /account/wishlist:
    get:
      operationId: listWishlistItems
      tags: [account]
      summary: List the authenticated customer's wishlist
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SavedProductReference" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: saveWishlistItem
      tags: [account]
      summary: Add or refresh an item in the authenticated customer's wishlist
      security: [{ ClerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductReferenceInput" }
      responses:
        "200":
          description: Saved wishlist item
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SavedProductReference" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: deleteWishlistItem
      tags: [account]
      summary: Remove an item from the authenticated customer's wishlist
      security: [{ ClerkSession: [] }]
      parameters:
        - name: shopifyProductId
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "204": { description: Wishlist item removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /account/recently-viewed:
    get:
      operationId: listRecentlyViewedItems
      tags: [account]
      summary: List the authenticated customer's recently viewed products
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Recently viewed items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ViewedProductReference" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: saveRecentlyViewedItem
      tags: [account]
      summary: Record a product view for the authenticated customer
      security: [{ ClerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductReferenceInput" }
      responses:
        "200":
          description: Recorded product view
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ViewedProductReference" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /account/orders:
    get:
      operationId: listCustomerOrders
      tags: [account]
      summary: Look up live Shopify orders using the authenticated customer's Clerk primary email
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Live Shopify order references
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrderReference" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: The authenticated Clerk primary email is unavailable or unverified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503":
          description: Shopify Admin order lookup is unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /account/orders/{orderId}:
    get:
      operationId: getCustomerOrder
      tags: [account]
      summary: Get one live Shopify order owned by the authenticated customer
      security: [{ ClerkSession: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9]+$", minLength: 1, maxLength: 32 }
      responses:
        "200":
          description: Live Shopify order details and fulfillment tracking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Order not found for the authenticated customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "422":
          description: The authenticated Clerk primary email is unavailable or unverified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "503":
          description: Shopify Admin order lookup is unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Confirm the current Clerk user has staff access
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Authorized staff session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/staff:
    get:
      operationId: listAdminStaff
      tags: [admin]
      summary: List users and their dashboard access for an owner
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Clerk users and current dashboard roles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminStaffMember" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/staff/{userId}:
    patch:
      operationId: updateAdminStaffAccess
      tags: [admin]
      summary: Grant or revoke staff access as an owner
      security: [{ ClerkSession: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminStaffAccessUpdate" }
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminStaffMember" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Clerk user not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/policies:
    get:
      operationId: listAdminPolicies
      tags: [admin]
      summary: List policy and sizing content
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Policy documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PolicyDocument" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/policies/{slug}:
    put:
      operationId: saveAdminPolicy
      tags: [admin]
      summary: Draft, submit for review, or publish policy and sizing content
      security: [{ ClerkSession: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, pattern: "^[a-z0-9-]+$", minLength: 1, maxLength: 80 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyDocumentInput" }
      responses:
        "200":
          description: Saved policy document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyDocument" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Summarize recorded analytics events without inferred metrics
      security: [{ ClerkSession: [] }]
      responses:
        "200":
          description: Recorded event summary and unavailable metrics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAnalytics" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    ClerkSession:
      type: apiKey
      in: cookie
      name: __session
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: Staff access required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
      required: [error, code]
...
    AdminSession:
      type: object
      properties:
        authorized: { type: boolean, const: true }
        role: { type: string, enum: [owner, admin, staff] }
        canManageStaff: { type: boolean }
      required: [authorized, role, canManageStaff]
    AdminStaffMember:
      type: object
      properties:
        userId: { type: string }
        email: { type: string, format: email }
        displayName: { type: ["string", "null"] }
        role:
          type: ["string", "null"]
          enum: [owner, admin, staff, null]
      required: [userId, email, displayName, role]
    AdminStaffAccessUpdate:
      type: object
      properties:
        role:
          type: ["string", "null"]
          enum: [admin, staff, null]
      required: [role]
      additionalProperties: false
...
    PolicyDocument:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        contentMarkdown: { type: string }
        status: { type: string, enum: [draft, review, published] }
        updatedAt: { type: string, format: date-time }
      required: [slug, title, contentMarkdown, status, updatedAt]
    PolicyDocumentInput:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        contentMarkdown: { type: string, maxLength: 50000 }
        status: { type: string, enum: [draft, review, published] }
      required: [title, contentMarkdown, status]
      additionalProperties: false
    AnalyticsEventCount:
      type: object
      properties:
        eventName: { type: string }
        count: { type: integer, minimum: 0 }
      required: [eventName, count]
    AdminAnalytics:
      type: object
      properties:
        recordedEvents: { type: integer, minimum: 0 }
        firstRecordedAt: { type: ["string", "null"], format: date-time }
        lastRecordedAt: { type: ["string", "null"], format: date-time }
        events:
          type: array
          items: { $ref: "#/components/schemas/AnalyticsEventCount" }
        unavailableMetrics:
          type: array
          items: { type: string }
      required: [recordedEvents, firstRecordedAt, lastRecordedAt, events, unavailableMetrics]
    CustomerProfile:
      type: object
      properties:
        displayName: { type: ["string", "null"] }
        preferredCurrency: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [displayName, preferredCurrency, createdAt, updatedAt]
    CustomerProfileInput:
      type: object
      properties:
        displayName: { type: ["string", "null"], maxLength: 100 }
        preferredCurrency:
          type: ["string", "null"]
          pattern: "^[A-Z]{3}$"
      additionalProperties: false
    ProductReferenceInput:
      type: object
      properties:
        shopifyProductId: { type: string, minLength: 1, maxLength: 255 }
        productHandle: { type: string, minLength: 1, maxLength: 255 }
      required: [shopifyProductId, productHandle]
      additionalProperties: false
    SavedProductReference:
      type: object
      properties:
        shopifyProductId: { type: string }
        productHandle: { type: string }
        createdAt: { type: string, format: date-time }
      required: [shopifyProductId, productHandle, createdAt]
    ViewedProductReference:
      type: object
      properties:
        shopifyProductId: { type: string }
        productHandle: { type: string }
        viewedAt: { type: string, format: date-time }
      required: [shopifyProductId, productHandle, viewedAt]
    OrderReference:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        processedAt: { type: string, format: date-time }
        displayFinancialStatus: { type: ["string", "null"] }
        displayFulfillmentStatus: { type: ["string", "null"] }
        totalPrice: { $ref: "#/components/schemas/Money" }
      required: [id, name, processedAt, displayFinancialStatus, displayFulfillmentStatus, totalPrice]
    OrderLineItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        variantTitle: { type: ["string", "null"] }
        quantity: { type: integer, minimum: 1 }
        price: { $ref: "#/components/schemas/Money" }
      required: [id, title, variantTitle, quantity, price]
    OrderTracking:
      type: object
      properties:
        company: { type: ["string", "null"] }
        number: { type: ["string", "null"] }
        url: { type: string, format: uri }
      required: [company, number, url]
    OrderFulfillment:
      type: object
      properties:
        id: { type: string }
        status: { type: string }
        tracking:
          type: array
          items: { $ref: "#/components/schemas/OrderTracking" }
      required: [id, status, tracking]
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/OrderReference"
        - type: object
          properties:
            lineItems:
              type: array
              items: { $ref: "#/components/schemas/OrderLineItem" }
            fulfillments:
              type: array
              items: { $ref: "#/components/schemas/OrderFulfillment" }
          required: [lineItems, fulfillments]
...
    Money:
      type: object
      properties:
        amount: { type: string }
        currencyCode: { type: string }
      required: [amount, currencyCode]
...