# lib/api-zod/src/generated/types/adminAnalytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsEventCount } from './analyticsEventCount';

export interface AdminAnalytics {
  /** @minimum 0 */
  recordedEvents: number;
  /** @nullable */
  firstRecordedAt: Date | null;
  /** @nullable */
  lastRecordedAt: Date | null;
  events: AnalyticsEventCount[];
  unavailableMetrics: string[];
}