# lib/api-zod/src/generated/types/adminStaffMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStaffMemberRole } from './adminStaffMemberRole';

export interface AdminStaffMember {
  userId: string;
  email: string;
  /** @nullable */
  displayName: string | null;
  /** @nullable */
  role: AdminStaffMemberRole;
}