# lib/api-zod/src/generated/types/analyticsEventCount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsEventCount {
  eventName: string;
  /** @minimum 0 */
  count: number;
}